requirements.txt
Flask==2.3.3
mongoengine==0.27.0
python-dotenv==1.0.0
flask-cors==4.0.0
pymongo==4.5.0
dnspython==2.4.2
requests==2.31.0 